Swap cross-origin iframe to cross-origin request
I've picked the creative rotation example at random to try out the
createWorklet function here. Unlike the previous commit where you
can do a straight swap between addModule and createWorklet, in this
commit I've removed the iframe completely. As we're relying on the
data being managed by the script-origin I've also had to move the
seedStorage function into the worklet so it is written into the
storage for the content producer instead of the publisher.

Whenever we request a worklet from a different origin we need to
set the headers to allow cors and enable cross origin worklets. As
I've removed the iframe, I've moved the controls of the demo into
the invoking page directly (using hbs partials). For the sake of
the demo I've added the global getContentProducerUrl function to
make it easier to check the domain of the content producer within
the JS files for getting the worklet and the ads.

Also, as the demo is no longer being styled within the iframe I've
copied the css styles to the publisher pages.